requirements.txt

azure-iot-device
google-cloud-pubsub
python-dotenv